switched to BigDecimal for money calculations, added unit more tests

+++ src/main/java/com/example/tool_rental/model/RentalAgreement.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -14,7 +15,7 @@ public class RentalAgreement {
     String toolCode; //Specified at checkout
     String toolType; //From tool info
     String toolBrand; //From tool info
-    Float dailyRentalCharge; //Amount per day, specified by the tool type.
+    BigDecimal dailyRentalCharge; //Amount per day, specified by the tool type.
 
 //    CheckoutRequest checkoutRequest;
     Integer rentalDays; //Specified at checkout
@@ -24,9 +25,9 @@ public class RentalAgreement {
     @DateTimeFormat(pattern = "MM/dd/yy")
     LocalDate dueDate; //Calculated from checkout date and rental days.
     Integer chargeDays; //Count of chargeable days, from day after checkout through and including due date, excluding “no charge” days as specified by the tool type.
-    Float preDiscountCharge; //Calculated as charge days X daily charge. Resulting total rounded half up to cents.
-    Float discountAmount; //calculated from discount % and pre-discount charge. Resulting amount rounded half up to cents.
-    Float finalCharge; //Calculated as pre-discount charge - discount amount.
+    BigDecimal preDiscountCharge; //Calculated as charge days X daily charge. Resulting total rounded half up to cents.
+    BigDecimal discountAmount; //calculated from discount % and pre-discount charge. Resulting amount rounded half up to cents.
+    BigDecimal finalCharge; //Calculated as pre-discount charge - discount amount.
 
     //TODO consider Currency data type
     //TODO override tostring or some other print method
